Changed script.js checkSize, checkCellExists, and checkCellIsEmpty
When calling placeShip(ship, []) I was passing the array around a bit heavily. At some point I was passing in coordinates.length and then in checkSize it was taking in a parameter that was intended to be an array and it was receiving .length already which was causing an error in my check system.