All existing tests are now passing
after troubleshooting the gameboardFacV1.js in script.js and now going back to gamebaordFactV1, I believe all issues are delt with. The process slimmed down the Gameboard() function, I was making the checks overly complicated and handling passing and return values poorly. For instance, checkSize had been expecting an array, yet in checkAll I was passing in coords.length instead of coords. also new Error('') returns true when checking a boolean value, so the check was not catching the false as I thought it should.

Now everything is pure (to my knowledge) and the checks are working as expected. All bugs have been fixed, and I uncommented all jest tests for Gameboard() and everything is going as planned. We are officially back on track. Now to write some more tests.