Changed shipFactory.js from starting the ship arr at 1 to starting the ship arr to 0
the reason I am making this change is for gameboard.receiveAttack. The board.placeShip will now give the ship an array, using the ship prop/method .setCoords. I use this coordinates array to then find call .hit() on the correct spot of the ship. This design choice is so that moving forward things may be more consistent.

this also triggered a massive rehaul of shipFactory.test.js. Where I had to change all tests for .hit(), and the expected outcome arrays. Had to change [1,2,3] [0,1,2]. I ran all tests and everything is clear, able to move forward again.