The fundamental game works!
the boards can be clicked on, the bot fires a shot off each time the player does. The game works.

next on the list is to move all the logic currently in script.js to other files as to clean up script.

And to establish a fundamental game loop, once the player hits start.