highlight and danger mouseover and mouseout are working as intended pre other ships being placed.
the bug I was running into was because I was stopping endPoint at 9, in the handleClick.js file (which I probably should rename later. Changing endpoint to 10 fixed the problem. I also made the code slightly more simple and removed all logs and comments.)

the handleClick.js file is very similar to the pickRandom logic. So the goal might be to condence both files into one, and just have random pick a random starting cell, and the handleClick.js logic will generate an array off of it.