highlight effect takes in to account non empty markers from the playerboard object
this is huge, as i'm moving back the core design which is that the view and logic are separate. The highlight - green or red depends on the board object itself if the items being selected for ship placement are available. INSTEAD of doing off the dom itself, I go off the BOARD OBJECT itself to render the proper visual aid (green = good, red = bad.) I even mark the exact spots where the ship would be invalid, unless the ship is off screen, then any cells highlight will be red.