starting to work on the end screen modal.
put some tests so I can trigger the end of the game quickly and easily, that way I can test the modal with one click. I'm sure I could find a way to do it without clicking, but this will work lol. Now I just need to set up the modal correctly.

the one time I reset the board, the player side was still using the board from the previous game, which is something I will need to fix.